Lesson 66

+++ MyCourse/Models/Entities/Lesson.cs
@@ -2,8 +2,8 @@
 {
     public partial class Lesson
     {
-        public long Id { get; private set; }
-        public long CourseId { get; private set; }
+        public int Id { get; private set; }
+        public int CourseId { get; private set; }
         public string Title { get; private set; }
         public string Description { get; private set; }
         public string Duration { get; private set; }
